GC bug：修复sweep阶段结束后，还会将gcobject标记为灰色的bug，导致下一轮gc中，无法被标记和扫描。

+++ C06/tetris/dummylua/test/p3_test.c
@@ -96,7 +96,7 @@ void test_gc(struct lua_State* L) {
 }
 
 void p3_test_main() {
-    printf("short string len:%lld, long string len:%lld \n", strlen(g_shrstr), strlen(g_lngstr));
+    printf("short string len:%lld, long string len:%lld \n", (long long)strlen(g_shrstr), (long long)strlen(g_lngstr));
 
     struct lua_State* L = luaL_newstate();
